chore(watch): inline bash module
We are only using this module in the watch module, so let’s get rid of
it. It’s just one function after all.
